# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo MiauChei",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Local"
        }
    ],
    "paths": {
...
        "/pet-found/{id}": {
            "put": {
                "tags": [
                    "Pets"
                ],
                "summary": "Pet encontrado",
                "description": "Atualização do pet para encontrado",
                "operationId": "a93ebbe4f3e3881ab92e4d4a22b6f563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recents": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Pets recente",
                "description": "Retorno de pets cadastrados recentemente",
                "operationId": "920d85d717fe45832624cedf5e7caf85",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mypets": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Meus Pets",
                "description": "Retorno dos pets cadastrados pelo usuário logado",
                "operationId": "dd8f2a06ba9a7706ba23f2517a13ab1e",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pets-lost": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Pets perdidos",
                "description": "Retorna lista de pets perdidos",
                "operationId": "124038c31a7407d99256ce4bd2a9ef19",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pets-details/{id}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Detalhes do pet",
                "description": "Retorna detalhes do pet conforme id passado",
                "operationId": "cf3c5e99e2d4583b9e69d201ec0bff4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pets-store": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Cadastrar pet",
                "description": "Retorna dados do pet",
                "operationId": "269baf138ad0ef622434350534dabaf2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "sex": {
                                        "type": "string"
                                    },
                                    "species": {
                                        "type": "string"
                                    },
                                    "breed": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "predominant_color": {
                                        "type": "string"
                                    },
                                    "secondary_color": {
                                        "type": "string"
                                    },
                                    "date_disappearance": {
                                        "type": "string"
                                    },
                                    "last_seen": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Thor",
                                    "sex": "M",
                                    "species": "Cachorro",
                                    "breed": "Vira Lata",
                                    "size": "Grande",
                                    "predominant_color": "Preto",
                                    "secondary_color": "Branco",
                                    "date_disappearance": "05/04/2022",
                                    "last_seen": "Rua teste",
                                    "photo": "teste123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}